Теперь содержимое списка обновляется в renderData
