--- app/src/main/java/ru/geekbrains/kotlinnotes/ui/MainActivity.kt
@@ -1,39 +1,0 @@
-package ru.geekbrains.kotlinnotes.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.GridLayoutManager
-import kotlinx.android.synthetic.main.activity_main.*
-import ru.geekbrains.kotlinnotes.R
-
-class MainActivity : AppCompatActivity() {
-
-    lateinit var viewModel: MainViewModel
-    lateinit var adapter: NotesRVAdapter
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        setSupportActionBar(toolbar)
-
-        rv_notes.layoutManager = GridLayoutManager(this, 2)
-        adapter = NotesRVAdapter { note ->
-            NoteEditActivity.start(this, note)
-        }
-        rv_notes.adapter = adapter
-
-        viewModel = ViewModelProvider(this).get(MainViewModel::class.java)
-        viewModel.getViewStateLiveData().observe(this, Observer {
-            it?.let {
-                adapter.notes = it.notes
-            }
-        })
-
-        fab_new_note.setOnClickListener {
-            NoteEditActivity.start(this)
-        }
-
-    }
-}
