+++ app/src/main/java/ru/geekbrains/kotlinnotes/ui/main/MainActivity.kt
@@ -1,0 +1,44 @@
+package ru.geekbrains.kotlinnotes.ui.main
+
+import android.os.Bundle
+import androidx.lifecycle.ViewModelProvider
+import androidx.recyclerview.widget.GridLayoutManager
+import kotlinx.android.synthetic.main.activity_main.*
+import ru.geekbrains.kotlinnotes.R
+import ru.geekbrains.kotlinnotes.data.entity.Note
+import ru.geekbrains.kotlinnotes.ui.NotesRVAdapter
+import ru.geekbrains.kotlinnotes.ui.base.BaseActivity
+import ru.geekbrains.kotlinnotes.ui.note.NoteEditActivity
+
+class MainActivity : BaseActivity<List<Note>?, MainViewState>() {
+
+    override val viewModel: MainViewModel by lazy {
+        ViewModelProvider(this).get(MainViewModel::class.java)
+    }
+
+    override val layoutRes = R.layout.activity_main
+    lateinit var adapter: NotesRVAdapter
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        setSupportActionBar(toolbar)
+
+        rv_notes.layoutManager = GridLayoutManager(this, 2)
+        adapter = NotesRVAdapter { note ->
+            NoteEditActivity.start(this, note.id)
+        }
+
+        rv_notes.adapter = adapter
+        fab_new_note.setOnClickListener {
+            NoteEditActivity.start(this)
+        }
+    }
+
+    override fun renderData(data: List<Note>?) {
+        data?.let {
+            adapter.notes = it
+        }
+    }
+
+}
